# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "digital-prescription",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/visits": {
      "get": {
        "tags": [
          "visits"
        ],
        "summary": "List Visits",
        "operationId": "list_visits_api_v1_visits_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "name": "doctor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doctor Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Patient name or phone number",
              "title": "Q"
            },
            "description": "Patient name or phone number"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, clinic timezone",
              "title": "Date From"
            },
            "description": "Inclusive, clinic timezone"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, clinic timezone",
              "title": "Date To"
            },
            "description": "Inclusive, clinic timezone"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "visits"
        ],
        "summary": "Create Visit",
        "operationId": "create_visit_api_v1_visits_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visits/{visit_id}": {
      "get": {
        "tags": [
          "visits"
        ],
        "summary": "Get Visit",
        "operationId": "get_visit_api_v1_visits__visit_id__get",
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BloodGroup": {
        "type": "string",
        "enum": [
          "O+",
          "O-",
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-"
        ],
        "title": "BloodGroup"
      },
...
      "DoctorBrief": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "specialization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization"
          },
          "qualification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualification"
          },
          "registration_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Number"
          },
          "accent_color": {
            "type": "string",
            "title": "Accent Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "specialization",
          "qualification",
          "registration_number",
          "accent_color"
        ],
        "title": "DoctorBrief",
        "description": "The doctor details embedded in patients, visits and printed prescriptions."
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "female",
          "male",
          "other"
        ],
        "title": "Gender"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "dosage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Dosage"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "duration": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MedicineCreate"
      },
      "MedicineRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dosage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dosage"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "dosage",
          "frequency",
          "duration",
          "instructions"
        ],
        "title": "MedicineRead"
      },
...
      "PatientBrief": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_number": {
            "type": "integer",
            "title": "Patient Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "blood_group": {
            "$ref": "#/components/schemas/BloodGroup"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_number",
          "full_name",
          "age",
          "gender",
          "phone",
          "blood_group",
          "allergies"
        ],
        "title": "PatientBrief",
        "description": "The patient details embedded in visits and printed prescriptions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisitCreate": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "vitals": {
            "$ref": "#/components/schemas/Vitals",
            "default": {}
          },
          "complaints": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Complaints"
          },
          "diagnosis": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Diagnosis"
          },
          "medicines": {
            "items": {
              "$ref": "#/components/schemas/MedicineCreate"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Medicines"
          },
          "advice": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Advice"
          },
          "follow_up_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "patient_id",
          "diagnosis",
          "medicines"
        ],
        "title": "VisitCreate",
        "description": "A new consultation. The prescribing doctor is always the logged-in user\nand the visit time is set by the server, so neither can be sent."
      },
      "VisitPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisitRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "patient_count": {
            "type": "integer",
            "title": "Patient Count"
          },
          "doctor_count": {
            "type": "integer",
            "title": "Doctor Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit",
          "patient_count",
          "doctor_count"
        ],
        "title": "VisitPage",
        "description": "A page of the visit log, plus totals across every matching visit."
      },
      "VisitRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "visited_at": {
            "type": "string",
            "format": "date-time",
            "title": "Visited At"
          },
          "vitals": {
            "$ref": "#/components/schemas/Vitals"
          },
          "complaints": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complaints"
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis"
          },
          "advice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advice"
          },
          "follow_up_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Date"
          },
          "medicines": {
            "items": {
              "$ref": "#/components/schemas/MedicineRead"
            },
            "type": "array",
            "title": "Medicines"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientBrief"
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorBrief"
          }
        },
        "type": "object",
        "required": [
          "id",
          "visited_at",
          "vitals",
          "complaints",
          "diagnosis",
          "advice",
          "follow_up_date",
          "medicines",
          "patient",
          "doctor"
        ],
        "title": "VisitRead"
      },
      "Vitals": {
        "properties": {
          "bp": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Bp"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "pulse": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Pulse"
          },
          "weight": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "spo2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Vitals",
        "description": "Free text, as doctors write them (\"120/80\", \"98.4\u00b0F\", \"62kg\"). All optional."
      }
    }
  }
}